Eased some eslint rules for spec-files
Disabled some eslint rules for testing. Type assertion and unsafe assertions are quite common, since typing-support of `.vue` files is not 100% accurate and some of the types often end in "any". Tests should then be built to fail if those properties are not as expected, and this has to be taken care of by the person writing the tests.